refactor: apply suggestions
